refactoring products API
