+++ cookwarestore/layers/ServiceLayer/products.js
@@ -1,0 +1,12 @@
+import { getProductsDAL } from "../DataAccessLayer/products";
+
+const getProductsSL = async () => {
+  try {
+    const products = await getProductsDAL();
+    return products;
+  } catch (err) {
+    return err;
+  }
+};
+
+export { getProductsSL };
